--- src/components/RightSideBar/BLock/Block.js
@@ -1,3 +1,0 @@
-import { Div } from './Block.styled';
-
-export const Block = ({ children }) => <Div>{children}</Div>;
