+++ src/components/RightSideBar/Container/Container.styled.jsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import backgroundUrl from 'static/images/Layer-4.png';
 
-export const MainContainer = styled.div`
+export const Container = styled.div`
   font-weight: 400;
   font-size: 14px;
   line-height: ${({ theme: { lineHeight } }) => lineHeight.verdanaHeight};
@@ -20,5 +20,23 @@ export const MainContainer = styled.div`
   @media (min-width: 1280px) {
     flex-direction: column;
     padding: 203px 0 0 106px;
+  }
+`;
+
+export const Wrapper = styled.div``;
+
+export const Div = styled.div`
+  &:first-child {
+    margin-bottom: 40px;
+
+    @media (min-width: 768px) {
+      margin-bottom: 0px;
+      margin-right: 97px;
+    }
+
+    @media (min-width: 1280px) {
+      margin-bottom: 60px;
+      margin-left: 0px;
+    }
   }
 `;
